requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
audioop-lts==0.2.2
audioread==3.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
decorator==5.2.1
fastapi==0.135.3
filelock==3.25.2
fsspec==2026.3.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
lazy-loader==0.5
librosa==0.11.0
llvmlite==0.47.0
MarkupSafe==3.0.3
more-itertools==11.0.2
mpmath==1.3.0
msgpack==1.1.2
networkx==3.6.1
numba==0.65.0
numpy==2.4.4
openai-whisper==20250625
packaging==26.0
pandas==3.0.2
platformdirs==4.9.6
pooch==1.9.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
regex==2026.4.4
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
setuptools==81.0.0
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.12.0
torch==2.11.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.44.0
